# backend/backend/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Loginoperations, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/Cart/addCart/{userId}/{productId}": {
      "post": {
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Cart/removeCart/{cartId}": {
      "delete": {
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "cartId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Cart/carts/{userId}": {
      "get": {
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Order/create/{userId}/{productId}": {
      "post": {
        "tags": [
          "Order"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Order/cancel/{orderId}": {
      "delete": {
        "tags": [
          "Order"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Order/updateOrderStatus": {
      "patch": {
        "tags": [
          "Order"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "orderId",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateStatusRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Order/{userId}": {
      "get": {
        "tags": [
          "Order"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductsResponse"
              }
            }
          }
        }
      }
    },
    "/api/v1/Product/add/{UserId}": {
      "post": {
        "tags": [
          "Product"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "UserId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddProductRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Product/delete/{id}": {
      "delete": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Product/Update/{productId}": {
      "put": {
        "tags": [
          "Product"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateProductRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Product/products": {
      "get": {
        "tags": [
          "Product"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductsResponse"
              }
            }
          }
        }
      }
    },
    "/api/v1/Product/products/{userId}": {
      "get": {
        "tags": [
          "Product"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductsResponse"
              }
            }
          }
        }
      }
    },
    "/api/v1/Product/get/{productId}": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/User/create": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateUserDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/User/login": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/User/forgot": {
      "put": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ForgotPasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/User/changePassword/{email}": {
      "put": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ChangePasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/{id}": {
      "delete": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "AddProductRequest": {
      "required": [
        "description",
        "location",
        "price",
        "title"
      ],
      "type": "object",
      "properties": {
        "title": {
          "minLength": 1,
          "type": "string"
        },
        "description": {
          "minLength": 1,
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "location": {
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "ChangePasswordRequest": {
      "required": [
        "activationCode",
        "newPassword"
      ],
      "type": "object",
      "properties": {
        "activationCode": {
          "minLength": 1,
          "type": "string"
        },
        "newPassword": {
          "maxLength": 256,
          "minLength": 8,
          "pattern": "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z]).{8,}$",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CreateUserDto": {
      "required": [
        "email",
        "password",
        "username"
      ],
      "type": "object",
      "properties": {
        "username": {
          "minLength": 1,
          "type": "string"
        },
        "email": {
          "format": "email",
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "maxLength": 256,
          "minLength": 8,
          "pattern": "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z]).{8,}$",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "ForgotPasswordRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "LoginRequest": {
      "required": [
        "email",
        "password"
      ],
      "type": "object",
      "properties": {
        "email": {
          "format": "email",
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "OrderStatus": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3
      ],
      "type": "integer"
    },
    "ProductsResponse": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "location": {
          "type": "string"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "user": {
          "$ref": "#/definitions/UserDto"
        },
        "datePosted": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UpdateProductRequest": {
      "required": [
        "description",
        "price",
        "title"
      ],
      "type": "object",
      "properties": {
        "title": {
          "minLength": 1,
          "type": "string"
        },
        "description": {
          "minLength": 1,
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "UpdateStatusRequest": {
      "type": "object",
      "properties": {
        "status": {
          "$ref": "#/definitions/OrderStatus"
        }
      },
      "additionalProperties": false
    },
    "UserDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "username": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}